+++ src/sll/include/sll/vm.h
@@ -82,4 +82,15 @@ __SLL_EXTERNAL __SLL_CHECK_OUTPUT sll_object_t* sll_execute_function(sll_integer
 
 
 
+/**
+ * \flags check_output func
+ * \name sll_get_call_stack
+ * \group vm
+ * \desc Docs!
+ * \ret const sll_call_stack_t*
+ */
+__SLL_EXTERNAL __SLL_CHECK_OUTPUT const sll_call_stack_t* sll_get_call_stack(void);
+
+
+
 #endif